feat: 패키지 구조 및 baseId 구현
