+++ be/src/main/java/com/foodymoody/be/notification/domain/NotificationId.java
@@ -1,32 +1,15 @@
 package com.foodymoody.be.notification.domain;
 
-import com.foodymoody.be.common.util.IdGenerator;
-import java.io.Serializable;
+import com.foodymoody.be.common.util.BaseId;
 import javax.persistence.Embeddable;
 import lombok.AccessLevel;
-import lombok.RequiredArgsConstructor;
+import lombok.NoArgsConstructor;
 
-@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Embeddable
-public class NotificationId implements Serializable {
+public class NotificationId extends BaseId {
 
-    private static final long serialVersionUID = -5351955236373496259L;
-
-    private String value;
-
-    private NotificationId(String value) {
-        this.value = value;
-    }
-
-    public static NotificationId from(String id) {
-        return new NotificationId(id);
-    }
-
-    public static NotificationId newId() {
-        return new NotificationId(IdGenerator.generate());
-    }
-
-    public String getValue() {
-        return value;
+    NotificationId(String value) {
+        super(value);
     }
 }
