+++ be/src/main/java/com/foodymoody/be/notification/infra/NotificationRepository.java
@@ -1,4 +1,4 @@
-package com.foodymoody.be.notification.repository;
+package com.foodymoody.be.notification.infra;
 
 import com.foodymoody.be.notification.domain.Notification;
 import com.foodymoody.be.notification.domain.NotificationId;
@@ -16,17 +16,23 @@ public interface NotificationRepository extends JpaRepository<Notification, Noti
     Slice<Notification> findAllByMemberId(String memberId, Pageable pageable);
 
     @Modifying
-    @Query("UPDATE Notification _notification SET _notification.isRead = :status, _notification.updatedAt = :updatedAt WHERE _notification.id IN :notificationIds AND _notification.memberId = :memberId")
+    @Query("UPDATE Notification _notification "
+            + "SET _notification.isRead = :status, _notification.updatedAt = :updatedAt "
+            + "WHERE _notification.id IN :notificationIds AND _notification.memberId = :memberId")
     void updateAllStatus(@Param("status") boolean status, @Param("memberId") String memberId,
             @Param("updatedAt") LocalDateTime updatedAt,
             @Param("notificationIds") List<NotificationId> notificationIds);
 
     @Modifying
-    @Query("UPDATE Notification _notification SET _notification.isDeleted = true , _notification.updatedAt = :updatedAt WHERE _notification.memberId = :memberId")
+    @Query("UPDATE Notification _notification "
+            + "SET _notification.isDeleted = true , _notification.updatedAt = :updatedAt "
+            + "WHERE _notification.memberId = :memberId")
     void deleteAllByMemberId(@Param("memberId") String memberId, @Param("updatedAt") LocalDateTime updatedAt);
 
     @Modifying
-    @Query("UPDATE Notification _notification SET _notification.isDeleted = true , _notification.updatedAt = :updatedAt WHERE _notification.id IN :notificationIds AND _notification.memberId = :memberId")
+    @Query("UPDATE Notification _notification "
+            + "SET _notification.isDeleted = true , _notification.updatedAt = :updatedAt "
+            + "WHERE _notification.id IN :notificationIds AND _notification.memberId = :memberId")
     void deleteAllByIdIn(@Param("notificationIds") List<NotificationId> notificationIds,
             @Param("updatedAt") LocalDateTime updatedAt, @Param("memberId") String memberId);
